requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
django-cors-headers==4.7.0






